+++ lib/transport.c
@@ -22,7 +22,7 @@ int varlink_transport_listen(VarlinkURI *uri, char **pathp) {
 
 _public_ int varlink_listen(const char *address, char **pathp) {
         _cleanup_(varlink_uri_freep) VarlinkURI *uri = NULL;
-        long r;
+        int r;
 
         r = varlink_uri_new(&uri, address, false);
         if (r < 0)
